Former-commit-id: 5f68c60a9bb45d32b5c34aef7600081da25c741f

+++ id/src/main/java/com/simprints/id/data/db/local/LocalDbManager.kt
@@ -29,7 +29,7 @@ interface LocalDbManager {
                                 projectId: String? = null,
                                 userId: String? = null,
                                 moduleId: String? = null,
-                                toSync: Boolean? = null): Long
+                                toSync: Boolean? = null): Int
 
     fun getPeopleFromLocal(personId: String? = null,
                            projectId: String? = null,
